Added engagement formulas to Channel

+++ src/prj5/sortByReach.java
@@ -6,7 +6,6 @@ public class sortByReach implements GenericCompare<Channel> {
 
     @Override
     public int compare(Channel o1, Channel o2, int start, int end) {
-        return o1.getReach(start, end) - o2.getReach(start, end);
+        return (int)(o1.getReach(start, end) - o2.getReach(start, end));
     }
-    
 }
